+++ src/modules/Opportunities/Module.php
@@ -33,6 +33,23 @@ class Module extends \MapasCulturais\Module{
         $app->registerJobType(new Jobs\PublishResult(Jobs\PublishResult::SLUG));
         $app->registerJobType(new Jobs\UpdateSummaryCaches(Jobs\UpdateSummaryCaches::SLUG));
 
+
+        // ajusta validação da área de interesse
+        $app->hook('entity(Opportunity).validationErrors', function(&$errors) use ($app){
+            /** @var Opportunity $this */
+            if(isset($errors['term-area'])) {
+                if($this->parent){
+                    unset($errors['term-area']);
+                } else {
+                    foreach($errors['term-area'] as &$termError) {
+                        if(strpos($termError, i::__('área de atuação')) !== false) {
+                            $termError = str_replace(i::__('área de atuação'), i::__('área de interesse'), $termError);
+                        }
+                    }
+                }
+            }
+        });
+
         // atualiza o cache dos resumos das fase de avaliação
         $app->hook("entity(Registration).sent:before", function() use ($app) {
             /** @var Registration $this */